requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
python-multipart>=0.0.9   # required for FastAPI file uploads
pydantic>=2.0.0

